+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## WIP
+
+### Added
+
+- Add filter animation delayed-fade
+
+### Changed
+
+- Filter animation fade does not stagger
+
 ## 3.6.5 (December 21, 2020)
 
 ### Fixed